use separate dex.scala file for Dex object definition
Makes both files easier to navigate and reason about
